% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtwclust-methods.R
\docType{methods}
\name{clusterSim}
\alias{clusterSim}
\alias{clusterSim,dtwclust-method}
\title{Cluster Similarity Matrix}
\usage{
\S4method{clusterSim}{dtwclust}(object, data = NULL, method = c("shadow",
  "centers"), symmetric = FALSE, ...)
}
\arguments{
\item{object, data, method, symmetric, ...}{See \code{\link[flexclust:clusterSim]{flexclust::clusterSim()}}.}
}
\description{
No longer supported directly in \pkg{dtwclust}.
}
\seealso{
\code{\link[flexclust:clusterSim]{flexclust::clusterSim()}}
}
\keyword{internal}
