% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{consistency_check}
\alias{consistency_check}
\title{Check consistency of provided object}
\usage{
consistency_check(obj, case)
}
\arguments{
\item{obj}{Object to check.}

\item{case}{One of \code{c("ts", "tslist", "vltslist", "window", "tsmat")}}
}
\value{
\code{NULL} invisibly
}
\description{
Used extensively internally. Exported just for detection in case of parallel computing.
}

