% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-dtwclust.R
\name{create_dtwclust}
\alias{create_dtwclust}
\title{Create formal \code{dtwclust} objects}
\usage{
create_dtwclust(..., override.family = TRUE)
}
\arguments{
\item{...}{Any valid slots of \linkS4class{dtwclust}.}

\item{override.family}{Attempt to substitute the default family with one that conforms to the
provided elements? See details.}
}
\value{
A \linkS4class{dtwclust} object.
}
\description{
\strong{Deprecated} helper function to manually create formal \linkS4class{dtwclust} objects
}
\details{
Please see the initialize section in \link{tsclusters-methods} for the new objects..
}
