% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtwclust-methods.R
\docType{methods}
\name{cvi}
\alias{cvi}
\alias{cvi,dtwclust-method}
\title{Cluster validity indices}
\usage{
cvi(a, b = NULL, type = "valid", ..., log.base = 2)

\S4method{cvi}{dtwclust}(a, b = NULL, type = "valid", ..., log.base = 2)
}
\arguments{
\item{a}{An object returned by the \code{\link{dtwclust}} function, or a vector that can be coerced to
integers which indicate the cluster memeberships.}

\item{b}{If needed, a vector that can be coerced to integers which indicate the cluster memeberships.}

\item{type}{Character vector indicating which indices are to be computed. See details.}

\item{...}{Arguments to pass to and from other methods.}

\item{log.base}{Base of the logarithm to be used in the calculation of VI.}
}
\value{
The chosen CVIs
}
\description{
Compute different cluster validity indices (CVIs) of a given cluster partition, using the
clustering distance measure if applicable.
}
\details{
Clustering is commonly considered to be an unsupervised procedure, so evaluating its performance
can be rather subjective. However, a great amount of effort has been invested in trying to standardize
cluster evaluation metrics by using cluster validity indices (CVIs).

CVIs can be classified as internal, external or relative depending on how they are computed.
Focusing on the first two, the crucial difference is that internal CVIs only consider the partitioned
data and try to define a measure of cluster purity, whereas external CVIs compare the obtained partition
to the correct one. Thus, external CVIs can only be used if the ground truth is known. Each index defines
their range of values and whether they are to be minimized or maximized. In many cases, these CVIs can be
used to evaluate the result of a clustering algorithm regardless of how the clustering works internally,
or how the partition came to be.

Knowing which CVI will work best cannot be determined a priori, so they should be tested for each
specific application. Usually, many CVIs are utilized and compared to each other, maybe using a majority
vote to decide on a final result. Furthermore, it should be noted that many CVIs perform additional
distance calculations when being computed, which can be very considerable if using DTW.

Note that, even though a fuzzy partition can be changed into a crisp one, making it compatible with many
of the existing CVIs, there are also fuzzy CVIs tailored specifically to fuzzy clustering, and these may
be more suitable in those situations, but have not been implemented here yet.

External CVIs (the first 4 are calculated via \code{\link[flexclust]{comPart}}):

\itemize{
  \item \code{"RI"}: Rand Index (to be maximized).
  \item \code{"ARI"}: Adjusted Rand Index (to be maximized).
  \item \code{"J"}: Jaccard Index (to be maximized).
  \item \code{"FM"}: Fowlkes-Mallows (to be maximized).
  \item \code{"VI"}: Variation of Information (to be minimized).
}

Internal CVIs:

\itemize{
  \item \code{"Sil"}: Silhouette index (to be maximized). A normalized summation-type index based on
  intracluster distance and the distance to nearest neighbors. This index essentially calculates (or
  re-uses, if already available) the whole distance matrix using the clustering distance measure, so if
  you were trying to avoid that in the first place, this may not be the best CVI for your application.
}
}
\references{
Arbelaitz, O., Gurrutxaga, I., Muguerza, J., Perez, J. M., & Perona, I. (2013). An extensive
comparative study of cluster validity indices. Pattern Recognition, 46(1), 243-256.
}

