% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DBA.R
\name{DBA}
\alias{DBA}
\title{DTW Barycenter Averaging}
\usage{
DBA(X, center = NULL, max.iter = 25, error.check = TRUE, trace = FALSE)
}
\arguments{
\item{X}{A data matrix where each row is a time series. Optionally, a list where each element is a time series.}

\item{center}{Optionally, a time series to use as reference. Defaults to a random series of \code{X} if \code{NULL}.}

\item{max.iter}{Maximum number of iterations allowed.}

\item{error.check}{Should inconsistencies in the data be checked?}

\item{trace}{If \code{TRUE}, the current iteration is printed to screen.}
}
\value{
The average time series.
}
\description{
See Petitjean 2011
}

