% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-Distmat.R
\name{Distmat-generics}
\alias{Distmat-generics}
\alias{[,Distmat,ANY,ANY,ANY-method}
\alias{[,Distmat,ANY,ANY,ANY}
\title{Generics for \code{Distmat}}
\usage{
\S4method{[}{Distmat,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \linkS4class{Distmat} object.}

\item{i}{Row indices.}

\item{j}{Column indices.}

\item{...}{Ignored.}

\item{drop}{Logical to drop dimensions after subsetting.}
}
\description{
Generics with methods for \linkS4class{Distmat}.
}
\details{
Accessing matrix elements with \verb{[]} first calculates the values if necessary.
}
\keyword{internal}
