% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-DistmatLowerTriangular.R
\docType{class}
\name{DistmatLowerTriangular-class}
\alias{DistmatLowerTriangular-class}
\alias{DistmatLowerTriangular}
\title{Distance matrix's lower triangular}
\description{
Reference class that is used internally for PAM centroids when \code{pam.precompute = TRUE} and
\code{pam.sparse = FALSE}. It contains \linkS4class{Distmat}.
}
\section{Fields}{

\describe{
\item{\code{distmat}}{The lower triangular.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(..., distmat, series, distance, control, error.check = TRUE)}}{Initialization based on needed parameters}
}}

\keyword{internal}
