% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtw-lb.R
\name{dtw_lb}
\alias{dtw_lb}
\title{DTW calculation guided by Lemire's lower bound}
\usage{
dtw_lb(x, y = NULL, window.size = NULL, norm = "L1", error.check = TRUE)
}
\arguments{
\item{x}{A matrix where rows are time series, or a list of time series.}

\item{y}{An object similar to \code{x}.}

\item{window.size}{The window size to use with the DTW calculation. \strong{See details}.}

\item{norm}{Pointwise distance. Either \code{L1} for Manhattan distance or \code{L2} for Euclidean.}

\item{error.check}{Should inconsistencies in the data be checked?}
}
\value{
The distance matrix with class \code{crossdist}.
}
\description{
Calculation of a distance matrix with the Dynamic Time Warp (DTW) distance guided by Lemire's lower bound
(LB).
}
\details{
This function first calculates an initial estimate of a distance matrix between two sets of time series
using Lemire's improved lower bound. Afterwards, it uses the estimate to calculate the true DTW distances
of \emph{only} the nearest neighbors for each series in \code{x}.

Because of the way the different functions being used here are implemented, there is a subtle but critical
mismatch in the way the window size is defined for DTW and the LB. The DTW calculation with
\code{\link[dtw]{dtw}} expects an \emph{even} \code{window.size} that represents the distance between the
diagonal and one of the edges of the window. The LB calculation expects an \emph{odd} window.size that
represents the whole window width to be used in the running max and min. The outcome of said running
functions are centered with respect to the window width.

Therefore, if, for example, the DTW is calculated with a window of 10, the corresponding LB should
be calculated with \code{2*10 + 1 = 21}.

The function takes care of this discrepancy if needed, but you should be careful if you are
testing things manually.
}

