% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtwclust-classes.R
\docType{class}
\name{dtwclust-class}
\alias{dtwclust-class}
\title{Class definition for \code{dtwclust}}
\description{
Formal S4 class to know how to handle data for plotting.
}
\details{
It contains the following specific slots:

\itemize{
  \item \code{type}: A string indicating one of the supported clustering types of \code{\link{dtwclust}}.
  \item \code{distance}: A string indicating one of the supported distances of \code{\link{dtwclust}}.
  \item \code{centroid}: A string indicating one of the supported centroids of \code{\link{dtwclust}}.
}

Additionally, the class inherits from \code{\link[flexclust]{kccasimple-class}}, so all related slots and
methods are also supported.
}

