% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtwclust.R
\name{dtwclust}
\alias{dtwclust}
\title{Time series clustering with DTW}
\usage{
dtwclust(data = NULL, type = "partitional", k = 2, method = "average",
  distance = "dtw_lb", centroid = "median", window.size = NULL,
  norm = "L1", dc, control = NULL, save.data = FALSE, seed = NULL,
  trace = F, ...)
}
\arguments{
\item{data}{Numeric matrix where each row is a time series.}

\item{type}{What type of clustering method to use, \code{partitional}, \code{hierarchical} or \code{tadpole}.}

\item{k}{Numer of desired clusters in partitional methods.}

\item{method}{Which linkage method to use in hierarchical methods.}

\item{distance}{One of the supported distance measurements (see details). It can also be the name of a
family to use with function \code{\link[flexclust]{kcca}} if \code{type == "partitional"}, or a supported
distance of \code{\link[proxy]{dist}} if \code{type == "hierarchical"}.}

\item{centroid}{Either a supported string (see details) or an appropriate function to calculate centroids
when using partitional methods.}

\item{window.size}{Window constraint for DTW and LB calculations.}

\item{norm}{Pointwise distance. Either \code{L1} for Manhattan distance or \code{L2} for Euclidean.}

\item{dc}{Cutoff distance for TADPole algorithm.}

\item{control}{Parameters for partitional clustering algorithm. See \code{\link[flexclust]{flexclustControl}}.}

\item{save.data}{Return a copy of the data in the returned object?}

\item{seed}{Random seed for reproducibility}

\item{trace}{Boolean flag. If true, more output regarding the progress is printed to screen.}

\item{...}{Additional arguments to pass to \code{\link[proxy]{dist}}.}
}
\value{
An object with formal class \code{\link{dtwclust-class}} if \code{type == "partitional" | "tadpole"}. Otherwise
an object with class \code{hclust} as returned by \code{\link[stats]{hclust}}.
}
\description{
This function uses the DTW distance and related lower bounds to cluster time series. For now, all series
must have equal length.
}
\author{
Alexis Sarda
}

