% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtwclust-classes.R
\docType{class}
\name{dtwclustFamily-class}
\alias{dtwclustFamily-class}
\alias{dtwclustFamily}
\title{Class definition for \code{dtwclustFamily}}
\description{
\strong{Deprecated} S4 class with a family of functions used in \code{\link[=dtwclust]{dtwclust()}}.
}
\details{
Please refer to \linkS4class{tsclustFamily} for the new class.
}
\section{Slots}{

\describe{
\item{\code{dist}}{The function to calculate the distance matrices.}

\item{\code{allcent}}{The function to calculate centroids on each iteration.}

\item{\code{cluster}}{The function used to assign a series to a cluster.}

\item{\code{preproc}}{The function used to preprocess the data (relevant for \code{\link[stats:predict]{stats::predict()}}).}
}}

