% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dtwclustTimings}
\alias{dtwclustTimings}
\title{Results of timing experiments}
\format{TBA}
\source{
Refer to the package vignette.
}
\description{
This is the list with data frames containing the results of the timing experiments vignette
included with \pkg{dtwclust}.
}
