% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHINY-utils.R
\name{explore__plot}
\alias{explore__plot}
\title{This helper will produce the plot in the Explore tab panel.}
\usage{
explore__plot(ids, df, series)
}
\arguments{
\item{ids}{A character with the expression that should yield the integer ids.}

\item{df}{The data frame with the melted time series.}

\item{series}{The list of time series.}
}
\description{
This helper will produce the plot in the Explore tab panel.
}
\keyword{internal}
