% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHINY-utils.R
\name{explore__tidy_series}
\alias{explore__tidy_series}
\title{This helper will create the data frame used to plot in the Explore tab panel}
\usage{
explore__tidy_series(series)
}
\arguments{
\item{series}{The list of time series}
}
\value{
A data frame
}
\description{
This helper will create the data frame used to plot in the Explore tab panel
}
\keyword{internal}
