% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHINY-interactive-clustering.R
\name{interactive_clustering}
\alias{interactive_clustering}
\title{A shiny app for interactive clustering}
\usage{
interactive_clustering(series, ...)
}
\arguments{
\item{series}{Time series in the formats accepted by \code{\link[=tsclust]{tsclust()}}.}

\item{...}{More arguments for \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\description{
Display a shiny user interface to do clustering based on the provided series.
}
