% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NCCc.R
\name{NCCc}
\alias{NCCc}
\title{Cross-correlation with coefficient normalization}
\usage{
NCCc(x, y)
}
\arguments{
\item{x}{A series.}

\item{y}{Another series.}
}
\value{
The cross-correlation sequence with length \code{length(x) + length(y) - 1}.
}
\description{
This function uses FFT to compute the cross-correlation sequence between two series. They need not be of
equal length.
}

