% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parallel_dtwclust}
\alias{parallel_dtwclust}
\title{Allow parallel computation with \pkg{dtwclust} functions}
\usage{
parallel_dtwclust(flag)
}
\arguments{
\item{flag}{\code{TRUE} to allow use of parallel backends or \code{FALSE} to prevent it.}
}
\value{
\code{flag} invisibly
}
\description{
Set an internal flag that is checked by the functions that support parallel computation and
determines if they should attempt to use a parallel backend if one is registered.
}

