% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHINY-utils.R
\name{parse_input}
\alias{parse_input}
\title{This helper will parse comma-separated key-value pairs}
\usage{
parse_input(input_text, into = "list")
}
\arguments{
\item{input_text}{A character.}

\item{into}{The name of the function that will contain the resulting values.}
}
\description{
This helper will parse comma-separated key-value pairs
}
\keyword{internal}
