% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dtwclust-plot.R
\docType{methods}
\name{plot-dtwclust}
\alias{plot,dtwclust,missing-method}
\alias{plot-dtwclust}
\title{Plot the result of \code{dtwclust}}
\usage{
\S4method{plot}{dtwclust,missing}(x, y, clus = seq_len(x@k), data = NULL,
  ...)
}
\arguments{
\item{x}{An object of class \code{\link{dtwclust-class}} as returned by \code{\link{dtwclust}}.}

\item{y}{Ignored.}

\item{clus}{Which clusters to plot.}

\item{data}{The data in the same format as it was provided to \code{\link{dtwclust}}.}

\item{...}{Further arguments to pass to \code{\link[ggplot2]{geom_line}} for the plotting of the
\emph{cluster centers}. Default values are: \code{linetype = "dashed", size = 1.5, colour = "black",
alpha = 0.5}.}
}
\description{
Plots the time series of each cluster along with the obtained centroid. It uses \code{ggplot2} plotting
system.
}
\details{
The flag \code{save.data} must be set to \code{TRUE} when running \code{\link{dtwclust}} to be able to
use this.

Optionally, you can manually provide the clustering result as well as the data in \code{data}.
}
\seealso{
\code{\link{dtwclust-class}}, \code{\link{dtwclust}}
}

