% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinterpolate.R
\name{reinterpolate}
\alias{reinterpolate}
\title{Wrapper for simple linear reinterpolation}
\usage{
reinterpolate(x, new.length, multivariate = FALSE, newLength)
}
\arguments{
\item{x}{Data to reinterpolate. Either a vector, a matrix/data.frame where each row is to be reinterpolated, or a list of
vectors.}

\item{new.length}{Desired length of the output series.}

\item{multivariate}{Is \code{x} a multivariate time series? It will be detected automatically if a list is provided in
\code{x}.}

\item{newLength}{Deprecated.}
}
\value{
Reinterpolated time series
}
\description{
This function is just a wrapper for the native function \code{\link[stats]{approx}} to do simple linear
reinterpolation.
}

