% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SBD.R
\name{SBD}
\alias{SBD}
\title{Shape-based distance}
\usage{
SBD(x, y)
}
\arguments{
\item{x}{A time series.}

\item{y}{Another time series.}
}
\value{
A list with: \itemize{
  \item \code{dist}: The distance between \code{x} and \code{y}.
  \item \code{yshift}: A shifted version of \code{y} so that it optimally mathces \code{x}.
}
}
\description{
See Paparrizos 2015.
}
\details{
This function works best if the inputs are \emph{z-normalized}. If not, at least they should have
corresponding amplitudes, since the values of the signal \strong{does} affect the outcome.

If \code{x} and \code{y} do \strong{not} have the same length, it would be best if the longer sequence is
provided in \code{y}, because it will be shifted to match \code{x}. Anything before the matching point is
discarded and the series is padded with trailing zeros as needed.
}

