% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-SemiSupervisedDtw.R
\docType{class}
\name{SemiSupervisedDtw-class}
\alias{SemiSupervisedDtw-class}
\alias{SemiSupervisedDtw}
\title{Helper for semi-supervised DTW clustering}
\description{
Helper for semi-supervised DTW clustering
}
\section{Fields}{

\describe{
\item{\code{xptr}}{External pointer (C++ class). See corresponding file in src/utils/}
}}

\section{Methods}{

\describe{
\item{\code{get_unseen_pair()}}{Get a pair that is not contained in the graph,
NULL means the graph is connected.}

\item{\code{initialize(max_size)}}{Initialization of C++ helper}

\item{\code{link(i, j, link_type)}}{Link indices i and j.
dont_know = -1, cannot_link = 0, must_link = 1.
Returns TRUE if graph is connected after insertion.}
}}

\keyword{internal}
