% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape-extraction.R
\name{shape_extraction}
\alias{shape_extraction}
\title{Shape average of several time series}
\usage{
shape_extraction(X, cz = NULL, znorm = FALSE)
}
\arguments{
\item{X}{Numeric matrix where each row is a time series, or a list of time series.}

\item{cz}{Center to use as basis. \emph{It will be z-normalized}. Function uses all \code{X}
if \code{cz = NULL}. Must be provided and different than a zero-line if time series in
\code{X} have different lengths.}

\item{znorm}{Logical flag. Should z-scores be calculated for \code{X} before processing?}
}
\value{
Centroid time series.
}
\description{
Time-series shape extraction based on optimal alignments as proposed by Papparizos and Gravano, 2015, for
the k-Shape clustering algorithm.
}
\details{
This works only if the series are \emph{z-normalized}, since the output will also have this normalization.
The resulting centroid will have the same length as \code{cz} if provided, or the same length as the time
series in \code{X} if \code{cz = NULL}. Therefore, in the latter case, all series of \code{X} must have
equal lengths.

This centroid computation is casted as an optimization problem called maximization of Rayleigh Quotient.
See the cited article for more details.
}
\examples{

# Sample data
data(uciCT)

# Normalize desired subset
X <- zscore(CharTraj[1:5])

# Obtain centroid series
C <- shape_extraction(X)

# Result
matplot(do.call(cbind, X),
        type = "l", col = 1:5)
points(C)

}
\references{
Paparrizos J and Gravano L (2015). ``k-Shape: Efficient and Accurate Clustering of Time Series.'' In \emph{Proceedings of the 2015
ACM SIGMOD International Conference on Management of Data}, series SIGMOD '15, pp. 1855-1870. ISBN 978-1-4503-2758-9, \url{
http://doi.org/10.1145/2723372.2737793}.
}
\seealso{
\code{\link{SBD}}, \code{\link{zscore}}
}

