% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtwclust-methods.R
\docType{methods}
\name{silIndex}
\alias{silIndex}
\alias{silIndex,dtwclust-method}
\title{Silhouette cluster validity index}
\usage{
silIndex(x)

\S4method{silIndex}{dtwclust}(x)
}
\arguments{
\item{x}{An object returned by the \code{\link{dtwclust}} function.}
}
\value{
The Sil index, which is to be maximized across different values of \code{k}.
}
\description{
Compute the Silhouette (Sil) cluster validity index (CVI) of a given cluster partition using the
chosen distance measure.
}
\details{
A normalized summation-type index based on intracluster distance and the distance to nearest neighbors.

This index essentially calculates (or re-uses, if available) the whole distance matrix using the
chosen distance measure, so if you were trying to avoid that in the first place, this may not be the
best CVI for your application.
}
\references{
Arbelaitz, O., Gurrutxaga, I., Muguerza, J., Perez, J. M., & Perona, I. (2013). An extensive
comparative study of cluster validity indices. Pattern Recognition, 46(1), 243-256.
}

