% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseDistmat.R
\docType{class}
\name{SparseDistmat-class}
\alias{SparseDistmat-class}
\alias{SparseDistmat}
\title{Sparse distance matrix}
\description{
Reference class that is used internally for PAM centroids when \code{pam.precompute = FALSE}. It
allows for mutable state.
}
\section{Fields}{

\describe{
\item{\code{series}}{Time series list.}

\item{\code{distmat}}{The sparse matrix.}

\item{\code{distfun}}{The distance function to calculate the distance matrix.}

\item{\code{symmetric}}{Logical indicating if the matrix is symmetric.}

\item{\code{existing_ids}}{Matrix with the indices of existing values within the matrix.}

\item{\code{dist_args}}{Arguments for the distance function.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(..., series, distance, control, dist_args)}}{Initialization based on needed parameters}
}}

