% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseDistmat.R
\docType{methods}
\name{SparseDistmat-generics}
\alias{SparseDistmat-generics}
\alias{show,SparseDistmat-method}
\alias{show,SparseDistmat}
\alias{[,SparseDistmat,ANY,ANY,ANY-method}
\alias{[,SparseDistmat,ANY,ANY,ANY}
\title{Generics for \code{SparseDistmat}}
\usage{
\S4method{show}{SparseDistmat}(object)

\S4method{[}{SparseDistmat,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{object}{A \linkS4class{SparseDistmat} object.}

\item{x}{A \linkS4class{SparseDistmat} object.}

\item{i}{Row indices.}

\item{j}{Column indices.}

\item{...}{Ignored.}

\item{drop}{Logical to drop dimensions after subsetting.}
}
\description{
Generics with methods for \linkS4class{SparseDistmat}.
}
\details{
Accessing matrix elements with \code{[]} first calculates the values if necessary.
}
