% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsclust-controls.R
\name{tsclust-controls}
\alias{tsclust-controls}
\alias{partitional_control}
\alias{hierarchical_control}
\alias{tsclust-controls}
\alias{fuzzy_control}
\alias{tsclust-controls}
\alias{tadpole_control}
\alias{tsclust-controls}
\alias{tsclust_args}
\alias{tsclust-controls}
\title{Control parameters for clusterings with \code{\link{tsclust}}}
\usage{
partitional_control(pam.precompute = TRUE, iter.max = 100L, nrep = 1L,
  symmetric = FALSE, packages = character(0L), distmat = NULL)

hierarchical_control(method = "average", symmetric = FALSE,
  packages = character(0L), distmat = NULL)

fuzzy_control(fuzziness = 2, iter.max = 100L, delta = 0.001,
  packages = character(0L))

tadpole_control(dc, window.size, lb = "lbk")

tsclust_args(preproc = list(), dist = list(), cent = list())
}
\description{
Control parameters
}
\details{
Partitional

Hierarchical

Fuzzy

TADPole

Function arguments
}
