% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsclust-classes.R
\docType{class}
\name{TSClusters-class}
\alias{TSClusters-class}
\alias{PartitionalTSClusters-class}
\alias{HierarchicalTSClusters-class}
\alias{FuzzyTSClusters-class}
\title{Class definition for \code{TSClusters} and derived classes}
\description{
Formal S4 classes for time-series clusters.
}
\details{
TBD

TBD2

TBD3

TBD4
}
