% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsclust-family.R
\docType{class}
\name{tsclustFamily-class}
\alias{tsclustFamily-class}
\title{Class definition for \code{tsclustFamily}}
\description{
Formal S4 class with a family of functions used in \code{\link[=tsclust]{tsclust()}}.
}
\details{
The custom implementations also handle parallelization.

Since the distance function makes use of \code{proxy}, it also supports any extra \code{\link[proxy:dist]{proxy::dist()}}
parameters in \code{...}.

The prototype includes the \code{cluster} function for partitional methods, as well as a pass-through
\code{preproc} function.

Note that symmetric optimizations in \code{dist} (see \link{tsclust-controls}) are applied automatically to
distances that are included with \pkg{dtwclust} when only one data argument is provided. Use two
to avoid the optimizations (see examples).
}
\section{Slots}{

\describe{
\item{\code{dist}}{The function to calculate the distance matrices.}

\item{\code{allcent}}{The function to calculate centroids on each iteration.}

\item{\code{cluster}}{The function used to assign a series to a cluster.}

\item{\code{preproc}}{The function used to preprocess the data (relevant for \code{\link[stats:predict]{stats::predict()}}).}
}}

\note{
This class is meant to group together the most relevant functions, but they are \strong{not} linked
with each other automatically. In other words, neither \code{dist} nor \code{allcent} apply \code{preproc}. They
essentially don't know of each other's existence.
}
\examples{

# The dist() function in tsclustFamily works like proxy::dist() but supports
# parallelization and optimized symmetric calculations. If you like, you can
# use the function more or less directly, but provide a control argument when
# creating the family.

\dontrun{
data(uciCT)
fam <- new("tsclustFamily", dist = "gak")

# This is done with symmetric optimizations
crossdist <- fam@dist(CharTraj, window.size = 18L)

# This is done without symmetric optimizations
crossdist <- fam@dist(CharTraj, CharTraj, window.size = 18L)

# For non-dtwclust distances, symmetric optimizations only apply with an appropriate control
fam <- new("tsclustFamily", dist = "Euclidean",
           control = partitional_control(symmetric = TRUE))
fam@dist(CharTraj[1L:5L])
}

# If you want the fuzzy family, use fuzzy = TRUE
ffam <- new("tsclustFamily", control = fuzzy_control(), fuzzy = TRUE)


}
