% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uciCT.R
\name{uciCT}
\alias{CharTraj}
\alias{CharTrajLabels}
\alias{uciCT}
\alias{ucict}
\title{Subset of character trajectories data set}
\format{A list with 100 elements. Each element is a time series. Labels included as factor vector.}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Character+Trajectories}
}
\description{
Subset: only 5 examples of X velocity. See details.
}
\details{
Quoting the source:

"Multiple, labelled samples of pen tip trajectories recorded whilst writing individual characters.
All samples are from the same writer, for the purposes of primitive extraction. Only characters with a
single pen-down segment were considered."

The subset included here (\code{CharTraj}) has only 5 examples of the X velocity for each character.
A vector with labels is also loaded in \code{CharTrajLabels}.
}

