% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Wrapper for z-normalization}
\usage{
zscore(x, ...)
}
\arguments{
\item{x}{Data to normalize. Either a vector or a list of vectors.}

\item{...}{Further arguments to pass to \code{\link[base]{scale}}.}
}
\value{
Normalized data.
}
\description{
Wrapper for function \code{\link[base]{scale}} that returns zeros instead of \code{NaN}. It also
supports a list of vectors.
}

