% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtwclust-methods.R
\name{as.matrix}
\alias{as.matrix}
\title{as.matrix}
\usage{
as.matrix(x, ...)
}
\arguments{
\item{x, ...}{See \code{\link[base]{as.matrix}}.}
}
\description{
\pkg{proxy} exported a non-generic \code{as.matrix} function. This is to re-export the base
version and add some coercion methods for \code{pairdist} and \code{crossdist}.
}
\seealso{
\code{\link[base]{as.matrix}}
}

