% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-as-methods.R
\name{as.matrix}
\alias{as.matrix}
\title{as.matrix}
\usage{
as.matrix(x, ...)
}
\arguments{
\item{x, ...}{See \code{\link[base:as.matrix]{base::as.matrix()}}.}
}
\description{
\pkg{proxy} exported a non-generic \code{as.matrix} function. This is to re-export the base version
and add coercion methods for \code{pairdist} and \code{crossdist}.
}
\seealso{
\code{\link[base:as.matrix]{base::as.matrix()}}
}
