% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-clusterings.R
\name{compare_clusterings_configs}
\alias{compare_clusterings_configs}
\title{Create configurations for \code{\link{compare_clusterings}}}
\usage{
compare_clusterings_configs(types = c("p", "h", "f", "t"), k = 2L,
  controls = NULL, preprocs = preproc_configs(),
  distances = distance_configs(), centroids = centroid_configs())
}
\arguments{
\item{types}{Clustering types. It must be one of (possibly abbreviated): partitional,
hierarchical, fuzzy, tadpole.}

\item{k}{A numeric vector with one or more elements specifying the number of clusters to test.}

\item{controls}{A list of \code{\link{tsclust-controls}}. \code{NULL} means defaults. See
details.}

\item{preprocs}{Preprocessing configurations. See details.}

\item{distances}{Distance configurations. See details.}

\item{centroids}{Centroid configurations. See details.}
}
\value{
A list for each clustering type, each of which includes a data frame with the computed and
merged configurations. Each data frame has an extra attribute \code{num.configs} specifying the
number of configurations.
}
\description{
Create clustering configurations.
}
\details{
This function is based on \code{\link[base]{expand.grid}} and \code{\link[base]{merge}}.

Preprocessing, distance and centroid configurations are specified with the helper functions in
\code{\link{config-helpers}}, but using the example in \code{\link{compare_clusterings}} as basis
might be easier to understand.

The controls list must have one nested list of controls for the desired clustering types. Each
nested list may be specified with the usual \code{\link{tsclust-controls}} functions. Again,
please refer to the example in \code{\link{compare_clusterings}}.
}
