% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtwclust-classes.R
\name{dtwclust-class}
\alias{dtwclust-class}
\title{Class definition for \code{dtwclust}}
\description{
Formal S4 class to know how to handle data for plotting.
}
\details{
The class inherits from \code{\link[flexclust]{kccasimple-class}}, so most related slots and
methods are also supported.
}
\section{Slots}{

\describe{
\item{\code{type}}{A string indicating one of the supported clustering types of \code{\link{dtwclust}}.}

\item{\code{distance}}{A string indicating the distance used with \code{\link{dtwclust}}.}

\item{\code{centroid}}{A string indicating the centroid used with \code{\link{dtwclust}}.}

\item{\code{preproc}}{A string indicating the preprocessing used with \code{\link{dtwclust}}.}

\item{\code{datalist}}{The provided data in the form of a list, where each element is a time series.}

\item{\code{proctime}}{Time during function execution, as measured by \code{\link[base]{proc.time}}.}
}}

