% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{dtwclust-deprecated}
\alias{dtwclust-deprecated}
\title{Deprecated functionality in \pkg{dtwclust}}
\description{
The following functions and/or classes are deprecated and will be eventually removed:
\itemize{
\item \code{\link[=dtwclust]{dtwclust()}}
\item \linkS4class{dtwclustFamily}
\item \linkS4class{dtwclustControl}
\item \linkS4class{dtwclust}
\item \link{dtwclust-methods}
\item \code{\link[=randIndex]{randIndex()}}
\item \code{\link[=clusterSim]{clusterSim()}}
\item \code{\link[=create_dtwclust]{create_dtwclust()}}
\item \code{\link[=compute_envelop]{compute_envelop()}}
}
}
