% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtwclust-methods.R
\docType{methods}
\name{dtwclust-methods}
\alias{dtwclust-methods}
\alias{plot,dtwclust,missing}
\alias{plot,dtwclust,missing-method}
\alias{plot.dtwclust}
\alias{predict,dtwclust}
\alias{predict,dtwclust-method}
\alias{predict.dtwclust}
\alias{show,dtwclust}
\alias{show,dtwclust-method}
\alias{update,dtwclust}
\alias{update,dtwclust-method}
\alias{update.dtwclust}
\title{Methods for \code{dtwclust}}
\usage{
\S4method{show}{dtwclust}(object)

\method{update}{dtwclust}(object, ..., evaluate = TRUE)

\S4method{update}{dtwclust}(object, ..., evaluate = TRUE)

\method{predict}{dtwclust}(object, newdata = NULL, ...)

\S4method{predict}{dtwclust}(object, newdata = NULL, ...)

\method{plot}{dtwclust}(x, y, ..., clus = seq_len(x@k), labs.arg = NULL,
  data = NULL, time = NULL, plot = TRUE, type = NULL)

\S4method{plot}{dtwclust,missing}(x, y, ..., clus = seq_len(x@k),
  labs.arg = NULL, data = NULL, time = NULL, plot = TRUE, type = NULL)
}
\arguments{
\item{object, x}{An object of class \code{\link{dtwclust-class}} as returned by
\code{\link{dtwclust}}.}

\item{...}{For \code{plot}, further arguments to pass to \code{\link[ggplot2]{geom_line}} for the
plotting of the \emph{cluster centroids}, or to \code{\link[stats]{plot.hclust}}. See details.
For \code{update}, any supported argument. Otherwise ignored.}

\item{evaluate}{Logical. Defaults to \code{TRUE} and evaluates the updated call, which will
result in a new \code{dtwclust} object. Otherwise, it returns the unevaluated call.}

\item{newdata}{New data to be assigned to a cluster. It can take any of the supported formats of
\code{\link{dtwclust}}. Note that for multivariate series, this means that it \strong{must} be
a list of matrices, even if the list has only one element.}

\item{y}{Ignored.}

\item{clus}{A numeric vector indicating which clusters to plot.}

\item{labs.arg}{Arguments to change the title and/or axis labels. See \code{\link[ggplot2]{labs}}
for more information}

\item{data}{Optionally, the data in the same format as it was provided to \code{\link{dtwclust}}.}

\item{time}{Optional values for the time axis. If series have different lengths, provide the time
values of the longest series.}

\item{plot}{Logical flag. You can set this to \code{FALSE} in case you want to save the ggplot
object without printing anything to screen}

\item{type}{What to plot. \code{NULL} means default. See details.}
}
\value{
The plot method returns a \code{gg} object (or \code{NULL} for dendrogram plot) invisibly.
}
\description{
Methods associated with \code{\link{dtwclust-class}} objects.
}
\details{
Supported generics from the \code{flexclust} package are: \code{\link[flexclust]{randIndex}} and
\code{\link[flexclust]{clusterSim}}.

All generics from package \pkg{clue} are also supported in order to use its functions.

Show method displays basic information from the clustering results.

The \code{update} method takes the original function call, replaces any provided argument and
optionally evaluates the call again. Use \code{evaluate = FALSE} if you want to get the
unevaluated call.

The \code{predict} generic can take the usual \code{newdata} argument and it returns the
cluster(s) to which the data belongs; if \code{NULL}, it simply returns the obtained cluster
indices. It preprocesses the data with the corresponding function if available.
}
\section{Plotting}{


  The plot method uses the \code{ggplot2} plotting system (see \code{\link[ggplot2]{ggplot}}).

  The default depends on whether a hierarchical method was used or not. In those cases, the
  dendrogram is plotted by default; you can pass any extra parameters to
  \code{\link[stats]{plot.hclust}} via \code{...}.

  Otherwise, the function plots the time series of each cluster along with the obtained centroid.
  The default values for cluster centroids are: \code{linetype = "dashed"}, \code{size = 1.5},
  \code{colour = "black"}, \code{alpha = 0.5}. You can change this by means of \code{...}.

  You can choose what to plot with the \code{type} parameter. Possible options are:

  \itemize{
    \item \code{"dendrogram"}: Only available for hierarchical clustering.
    \item \code{"series"}: Plot the time series divided into clusters without including centroids.
    \item \code{"centroids"}: Plot the obtained centroids only.
    \item \code{"sc"}: Plot both series and centroids
  }

  The flag \code{save.data} should be set to \code{TRUE} when running \code{\link{dtwclust}} to be
  able to use this. Optionally, you can manually provide the data in the \code{data} parameter.

  If created, the function returns the \code{gg} object invisibly, in case you want to modify it to
  your liking. You might want to look at \code{\link[ggplot2]{ggplot_build}} if that's the case.

  If you want to free the scale of the X axis, you can do the following:

  \code{plot(object, plot = FALSE)} \code{+} \code{facet_wrap(~cl, scales = "free")}
}
\seealso{
\code{\link{dtwclust-class}}, \code{\link{dtwclust}}, \code{\link[ggplot2]{ggplot}},
\code{\link{cvi}}
}

