% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dtwclustTimings}
\alias{dtwclustTimings}
\title{Results of timing experiments}
\format{The results are organized into different data frames and saved in one list with nested lists.
For more details, refer to the included vignette or the scripts available at
\url{https://github.com/asardaes/dtwclust/tree/master/timing-experiments}.}
\source{
Refer to the timing experiments vignette.
}
\description{
This is the list with data frames containing the results of the timing experiments vignette
included with \pkg{dtwclust}. See \code{browseVignettes("dtwclust")}.
}
