% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lb-improved.R
\name{lb_improved}
\alias{lb_improved}
\title{Lemire's improved DTW lower bound}
\usage{
lb_improved(x, y, window.size, norm = "L1")
}
\arguments{
\item{x}{A time series.}

\item{y}{A time series with the same length as \code{x}.}

\item{window.size}{Window size for envelop calculation. \strong{See details}.}

\item{norm}{Pointwise distance. Either \code{L1} for Manhattan distance or \code{L2} for Euclidean.}
}
\value{
The improved lower bound for the DTW distance.
}
\description{
This function calculates a lower bound (LB) on the Dynamic Time Warp (DTW) distance between two time
series. It uses a Sakoe-Chiba constraint.
}
\details{
The lower bound is defined for time series of equal length only.

Because of the way the different functions being used here are implemented, there is a subtle but critical
mismatch in the way the window size is defined for DTW and the LB. The LB calculation expects an \emph{odd}
\code{window.size} that represents the whole window width to be used in the running max and min. The
outcome of said running functions are centered with respect to the window. The DTW calculation with
\code{\link[dtw]{dtw}} expects a window.size that represents the distance between the diagonal and one of
the edges of the window.

Therefore, if, for example, the LB is calculated with a window of 21, the corresponding DTW distance should
be calculated with \code{21 \%/\% 2 = 10}.

This function expects the \code{window.size} for the running max/min.
}
\examples{
# Sample data
data(uciCT)

# Lower bound distance between two series
d.lbi <- lb_improved(CharTraj[[1]], CharTraj[[2]], window.size = 11)

# Corresponding true DTW distance (accounting for window.size discrepancy)
d.dtw <- dtw(CharTraj[[1]], CharTraj[[2]], window.type = "slantedband", window.size = 5)$distance

d.lbi <= d.dtw
}
\references{
Lemire D (2009). ``Faster retrieval with a two-pass dynamic-time-warping lower bound .'' \emph{Pattern Recognition}, \strong{42}(9), pp.
2169 - 2180. ISSN 0031-3203, \url{http://doi.org/http://dx.doi.org/10.1016/j.patcog.2008.11.030}, \url{
http://www.sciencedirect.com/science/article/pii/S0031320308004925}.
}

