% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lb-keogh.R
\name{lb_keogh}
\alias{lb_keogh}
\title{Keogh's DTW lower bound}
\usage{
lb_keogh(x, y, window.size, norm = "L1")
}
\arguments{
\item{x}{A time series.}

\item{y}{A time series with the same length as \code{x}.}

\item{window.size}{Window size for envelop calculation. \strong{See details}.}

\item{norm}{Pointwise distance. Either \code{L1} for Manhattan distance or \code{L2} for Euclidean.}
}
\value{
A list with: \itemize{
  \item \code{d}: The lower bound of the DTW distance.
  \item \code{upper.env}: The time series of the upper envelope.
  \item \code{lower.env}: The time series of the lower envelope.
}
}
\description{
This function calculates a lower bound (LB) on the Dynamic Time Warp (DTW) distance between two time
series. It uses a Sakoe-Chiba constraint.
}
\details{
The lower bound is defined for time series of equal length only.

Because of the way the different functions being used here are implemented, there is a subtle but critical
mismatch in the way the window size is defined for DTW and the LB. The LB calculation expects an \emph{odd}
\code{window.size} that represents the whole window width to be used in the running max and min. The
outcome of said running functions are centered with respect to the window. The DTW calculation with
\code{\link[dtw]{dtw}} expects a window.size that represents the distance between the diagonal and one of
the edges of the window.

Therefore, if, for example, the LB is calculated with a window of 21, the corresponding DTW distance should
be calculated with \code{21 \%/\% 2 = 10}.

This function expects the \code{window.size} for the running max/min.
}
\examples{
# Sample data
data(uciCT)

# Lower bound distance between two series
d.lbk <- lb_keogh(CharTraj[[1]], CharTraj[[2]], window.size = 11)$d

# Corresponding true DTW distance (accounting for window.size discrepancy)
d.dtw <- dtw(CharTraj[[1]], CharTraj[[2]], window.type = "slantedband", window.size = 5)$distance

d.lbk <= d.dtw
}
\references{
Keogh E and Ratanamahatana CA (2005). ``Exact indexing of dynamic time warping.'' \emph{Knowledge and information systems}, \strong{7}(3),
pp. 358-386.
}

