% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb-keogh.R
\name{lb_keogh}
\alias{lb_keogh}
\title{Keogh's DTW lower bound}
\usage{
lb_keogh(x, y, window.size = NULL, norm = "L1", lower.env = NULL,
  upper.env = NULL)
}
\arguments{
\item{x}{A time series (reference).}

\item{y}{A time series with the same length as \code{x} (query).}

\item{window.size}{Window size for envelope calculation. See details.}

\item{norm}{Pointwise distance. Either \code{L1} for Manhattan distance or \code{L2} for Euclidean.}

\item{lower.env}{Optionally, a pre-computed lower envelope for \strong{\code{y}} can be provided.}

\item{upper.env}{Optionally, a pre-computed upper envelope for \strong{\code{y}} can be provided.}
}
\value{
A list with: \itemize{
  \item \code{d}: The lower bound of the DTW distance.
  \item \code{upper.env}: The time series of \code{y}'s upper envelope.
  \item \code{lower.env}: The time series of \code{y}'s lower envelope.
}
}
\description{
This function calculates a lower bound (LB) on the Dynamic Time Warp (DTW) distance between two time
series. It uses a Sakoe-Chiba constraint.
}
\details{
The windowing constraint uses a centered window. The calculations expect a value in \code{window.size}
that represents the distance between the point considered and one of the edges of the window. Therefore,
if, for example, \code{window.size = 10}, the warping for an observation \eqn{x_i} considers the points
between \eqn{x_{i-10}} and \eqn{x_{i+10}}, resulting in \code{10(2) + 1 = 21} observations falling within
the window.

The reference time series should go in \code{x}, whereas the query time series should go in \code{y}.
}
\note{
The lower bound is defined for time series of equal length only and is \strong{not} symmetric.

If you wish to calculate the lower bound between several time series, it would be better to use the version
registered with the \code{proxy} package, since it includes some small optimizations. See the examples.

However, because of said optimizations and the way \code{proxy}'s \code{\link[proxy]{dist}} works, the
latter's \code{pairwise} argument will not work with this distance. You can use the custom argument
\code{force.pairwise} to get the correct result.
}
\examples{

# Sample data
data(uciCT)

# Lower bound distance between two series
d.lbk <- lb_keogh(CharTraj[[1]], CharTraj[[2]], window.size = 20)$d

# Corresponding true DTW distance
d.dtw <- dtw(CharTraj[[1]], CharTraj[[2]],
             window.type = "slantedband", window.size = 20)$distance

d.lbk <= d.dtw

# Calculating the LB between several time series using the 'proxy' package
# (notice how both argments must be lists)
D.lbk <- proxy::dist(CharTraj[1], CharTraj[2:5], method = "LB_Keogh",
                     window.size = 20, norm = "L2")

# Corresponding true DTW distance
# (see dtwclust-package description for an explanation of DTW2)
D.dtw <- proxy::dist(CharTraj[1], CharTraj[2:5], method = "DTW2",
                     window.type = "slantedband", window.size = 20)

D.lbk <= D.dtw

}
\references{
Keogh E and Ratanamahatana CA (2005). ``Exact indexing of dynamic time warping.'' \emph{Knowledge
and information systems}, \strong{7}(3), pp. 358-386.
}

