% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NCCc.R
\name{NCCc}
\alias{NCCc}
\title{Cross-correlation with coefficient normalization}
\usage{
NCCc(x, y)
}
\arguments{
\item{x}{A time series.}

\item{y}{Another time series.}
}
\value{
The cross-correlation sequence with length \code{length(x) + length(y) - 1}.
}
\description{
This function uses FFT to compute the cross-correlation sequence between two series. They need not be of
equal length.
}
\references{
Paparrizos J and Gravano L (2015). ``k-Shape: Efficient and Accurate Clustering of Time Series.'' In \emph{Proceedings of the 2015
ACM SIGMOD International Conference on Management of Data}, series SIGMOD '15, pp. 1855-1870. ISBN 978-1-4503-2758-9, \url{
http://doi.org/10.1145/2723372.2737793}.
}
\seealso{
\code{\link{SBD}}
}

