% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinterpolate.R
\name{reinterpolate}
\alias{reinterpolate}
\title{Wrapper for simple linear reinterpolation}
\usage{
reinterpolate(ts, newLength)
}
\arguments{
\item{ts}{A time series.}

\item{newLength}{Desired length of the output series.}
}
\value{
Reinterpolated time series
}
\description{
This function is just a wrapper for the native function \code{\link[stats]{approx}} to do simple linear reinterpolation.
}

