% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shape-extraction.R
\name{shape_extraction}
\alias{shape_extraction}
\title{Shape average of several time series}
\usage{
shape_extraction(X, cz = NULL, znorm = FALSE)
}
\arguments{
\item{X}{Numeric matrix where each row is a time series.}

\item{cz}{Center to use as basis. Should already be \emph{normalized}. Calculation uses all \code{X}
if \code{cz = NULL}.}

\item{znorm}{Boolean flag. Should z-scores be calculated for \code{X} before processing?}
}
\value{
Centroid time series.
}
\description{
See Paparrizos 2015.
}
\details{
This works best (perhaps only) if the signals have similar amplitudes, which is why \emph{z-normalization}
is recommended.
}

