% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-SparseDistmat.R
\docType{class}
\name{SparseDistmat-class}
\alias{SparseDistmat-class}
\alias{SparseDistmat}
\title{Sparse distance matrix}
\description{
Reference class that is used internally for PAM centroids when \code{pam.precompute = FALSE} and
\code{pam.sparse = TRUE}. It allows for mutable state. It contains \linkS4class{Distmat}.
}
\section{Fields}{

\describe{
\item{\code{distmat}}{The sparse matrix.}

\item{\code{symmetric}}{Logical indicating if the matrix is symmetric.}

\item{\code{distmat_indices}}{External pointer (C++ class) with the indices of existing values within
the matrix, and the method to update them.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(..., distmat, series, distance, control, error.check = TRUE)}}{Initialization based on needed parameters}
}}

\keyword{internal}
