% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TADPole.R
\name{TADPole}
\alias{TADPole}
\title{TADPole clustering}
\usage{
TADPole(data, window.size = NULL, k = 2, dc, error.check = TRUE)
}
\arguments{
\item{data}{The data matrix. Optionally a list with each time series.}

\item{window.size}{Window size constraint for DTW.}

\item{k}{The number of desired clusters.}

\item{dc}{The cutoff distance.}

\item{error.check}{Should inconsistencies in the data be checked?}
}
\value{
A list with: \itemize{
  \item \code{cl}: Cluster indices.
  \item \code{centers}: Indices of the centers.
  \item \code{distCalcPercentage}: Percentage of distance calculations that were actually performed.
}
}
\description{
See Begum et al. 2015
}

