% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-data.R
\name{uciCT}
\alias{uciCT}
\alias{ucict}
\alias{CharTraj}
\alias{CharTrajLabels}
\alias{CharTrajMV}
\title{Subset of character trajectories data set}
\format{Lists with 100 elements. Each element is a time series. Labels included as factor vector.}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Character+Trajectories}
}
\description{
Subset: only 5 examples of each considered character. See details.
}
\details{
Quoting the source:

"Multiple, labelled samples of pen tip trajectories recorded whilst writing individual
characters. All samples are from the same writer, for the purposes of primitive extraction. Only
characters with a single pen-down segment were considered."

The subset included in \code{CharTraj} has only 5 examples of the X velocity for each character. A
vector with labels is also loaded in \code{CharTrajLabels}.

The subset included in \code{CharTrajMV} has 5 examples too, but includes tip force as well as X and Y
velocity. Each element of the list is a multivariate series with 3 variables.

Please note that even though both \code{CharTraj} and \code{CharTrajMV} have the same series names, the
actual series in each subset are \strong{not} the same, i.e., \code{CharTraj$A.V1} is not in
\code{CharTrajMV$A.V1}.
}
